se agregan validaciones en los services

+++ src/main/java/com/ar/conde/reservaDeTurnos/service/OdontologoService.java
@@ -1,18 +1,17 @@
 package com.ar.conde.reservaDeTurnos.service;
 
 import com.ar.conde.reservaDeTurnos.entity.Odontologo;
+import com.ar.conde.reservaDeTurnos.log4j.Log4j;
 import com.ar.conde.reservaDeTurnos.repositories.OdontologoRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service("odontologo")
 public class OdontologoService implements IService<Odontologo>{
-
-
 
     private OdontologoRepository repository;
 
@@ -21,15 +20,28 @@ public class OdontologoService implements IService<Odontologo>{
     }
 
     @Override
-    @Transactional(readOnly = true)
+    @Transactional()
     public List getAll() {
-        return (List<Odontologo>) repository.findAll();
+
+        try{
+            return (List<Odontologo>) repository.findAll();
+        } catch (Exception e){
+            Log4j.error(e.toString());
+            throw e;
+        }
+
     }
 
     @Override
-    @Transactional(readOnly = true)
+    @Transactional()
     public Optional<Odontologo> getById(Long id) {
-        return repository.findById(id);
+        try{
+            return repository.findById(id);
+        } catch (Exception e){
+            Log4j.error(e.toString());
+            throw e;
+        }
+
     }
 
     @Override
@@ -38,12 +50,14 @@ public class OdontologoService implements IService<Odontologo>{
        try{
            Optional<Odontologo> isexists = repository.buscarPorMatricula(odontologo.getMatricula());
            if(isexists.isPresent()){
-                throw new IllegalArgumentException("la Matricula ya esta en uso");
+                throw new IllegalArgumentException("la Matricula ya esta registrada ");
+
            }
            return repository.save(odontologo);
        }
-       catch(Exception exception){
-           throw exception;
+       catch(Exception e){
+           Log4j.error(e.toString());
+           throw e;
        }
 
     }
@@ -51,13 +65,49 @@ public class OdontologoService implements IService<Odontologo>{
     @Override
     @Transactional
     public Odontologo update(Long id, Odontologo odontologo) {
-        return null;
+
+        try{
+            Optional<Odontologo> isExists = getById(id);
+            if(isExists.isEmpty()){
+                throw new IllegalArgumentException("el ID ingresado no existe ");
+            }
+
+            Optional<Odontologo> odontologoExist = repository.buscarPorMatricula(odontologo.getMatricula());
+            if(odontologoExist.isPresent() && !Objects.equals(odontologoExist.get().getMatricula(), odontologo.getMatricula())){
+                throw new IllegalArgumentException("la matricula ya existe ");
+            }
+
+                Odontologo odontologoUpdate = isExists.get();
+                odontologoUpdate.setApellido(odontologo.getApellido());
+                odontologoUpdate.setNombre(odontologo.getNombre());
+                odontologoUpdate.setMatricula(odontologo.getMatricula());
+
+                return repository.save(odontologoUpdate);
+        }
+        catch (Exception e){
+            Log4j.error(e.toString());
+            throw e;
+
+        }
     }
 
 
     @Override
     @Transactional
     public void delete(Long id) {
-        repository.deleteById(id);
+
+        try{
+            Optional<Odontologo> odontologo = getById(id);
+            if(odontologo.isPresent()){
+                repository.deleteById(id);
+            } else {
+                throw new IllegalArgumentException("El id " + id +" no existe");
+            }
+
+        } catch (Exception e){
+            Log4j.error(e.toString());
+            throw e;
+        }
+
     }
 }
